+++ internal/bundler_tests/bundler_ts_test.go
@@ -1054,6 +1054,7 @@ func TestTSExperimentalDecorators(t *testing.T) {
 					@x @y mDef = 1
 					@x @y method(@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
 					@x @y declare mDecl
+					@x @y abstract mAbst
 					constructor(@x0 @y0 arg0, @x1 @y1 arg1) {}
 
 					@x @y static sUndef
@@ -1070,6 +1071,7 @@ func TestTSExperimentalDecorators(t *testing.T) {
 					@x @y [mDef()] = 1
 					@x @y [method()](@x0 @y0 arg0, @x1 @y1 arg1) { return new Foo }
 					@x @y declare [mDecl()]
+					@x @y abstract [mAbst()]
 
 					// Side effect order must be preserved even for fields without decorators
 					[xUndef()]